census: one definition of which files can invoke something (#1213)
`tests/digest/no-host-digest-tools.mjs` landed in #1607 with its own copy of
the census's tree walk -- `git ls-files`, skip the fixture directory, keep
what has a runnable extension or a shebang -- and the copy was wrong inside the
week. It had dropped this branch:

    if (!named && body.indexOf(String.fromCharCode(0)) !== -1) continue

which `check.mjs` carries with a measurement attached: a NUL rules out a
candidate found by shebang sniffing, but not a file whose extension already
says it is a script, because `tests/interop/bindgen-c/check-report.mjs` embeds
four NUL bytes in a fixture string and skipping it dropped a real Node.js
source file from the census. Two scanners asking the same question in the same
words had already stopped agreeing about which files the tree contains.

So the walk moves to `tooling/forbidden-requirements/universe.mjs` and both
callers import it. The digest gate loses 40 lines and gains the NUL rule;
`check.mjs` loses the definition and keeps everything else.

Not `detect.mjs`, which would have been the shorter import: its header says it
is separated from the checker so the self-test can drive the detectors over
fixtures without scanning the repository, and a module that shells out to `git
ls-files` on import takes that property away. The prose form of the set stays
there as `FILE_SET`, beside the per-detector predicates and matching its
sibling in `tooling/machine-dependent/`, because `--predicates` prints them
together; `universe.mjs` says where it went and why.

Two related duplications are deliberately left alone. `tests/rfc/validate-
registry.mjs` and `tests/release/validate-claims.mjs` each define a three-line
`trackedFiles()`, but neither wants this file set -- no shebang sniffing, no
fixture exclusion -- and a shared name that means a different set in each caller
is worse than three short functions. `tooling/machine-dependent/` has its own
`FILE_SET` over `*.sh *.mjs *.js` for the same reason: it is a different
question, correctly asked separately.

The scan's own predicate is unchanged, and both directions of the gate still
report: 5 invocation shapes caught, 5 mentions refused, and the tree clean --
438 files now rather than 437, which is `universe.mjs` itself entering the set
it defines.

Validation: task digest PASS, the census checker PASS on 675 rows with the new
`node source` row for `universe.mjs`, its self-test PASS including the fixture-
exclusion assertion that keeps the one hole from widening, task
gate-reachability PASS, and a full local task verify.
