+++ tooling/forbidden-requirements/check.mjs
@@ -33,13 +33,13 @@
  * grow.
  */
 
-import { execFileSync } from 'node:child_process'
-import { readFileSync, statSync } from 'node:fs'
+import { readFileSync } from 'node:fs'
 import { dirname, join, resolve } from 'node:path'
 import { fileURLToPath } from 'node:url'
 
 import { DETECTORS, FILE_SET, dialectOf, withoutComments } from './detect.mjs'
 import { NEEDS, needOf, reachability } from './reach.mjs'
+import { trackedFiles, universe } from './universe.mjs'
 
 const HERE = dirname(fileURLToPath(import.meta.url))
 const ROOT = resolve(HERE, '..', '..')
@@ -51,14 +51,6 @@ const ROOT = resolve(HERE, '..', '..')
  */
 const LEDGER = process.env.KOFUN_FORBIDDEN_CENSUS ?? join(HERE, 'census.tsv')
 const CONTRACT = join(ROOT, 'spec', 'native-toolchain-v1', 'contract.json')
-
-/*
- * The fixture directory is the one place excluded from the scan, because it
- * contains a deliberate invocation whose whole purpose is to be detected by the
- * self-test. Excluding it is stated here and asserted in the self-test, so the
- * hole cannot quietly widen into "anything under tooling/".
- */
-const EXCLUDED = 'tooling/forbidden-requirements/fixtures/'
 
 /*
  * The file that spells the patterns out is exempt from `invoke` detection, and
@@ -83,47 +75,6 @@ const CLASSES = new Set(['required-today', 'removable'])
 const fail = (message) => {
     process.stderr.write(`FAIL: forbidden requirements: ${message}\n`)
     process.exitCode = 1
-}
-
-/* ------------------------------------------------------------------ the tree */
-
-function trackedFiles() {
-    const out = execFileSync('git', ['ls-files', '-z'], { cwd: ROOT, maxBuffer: 64 * 1024 * 1024 })
-    return out.toString('utf8').split('\0').filter(Boolean)
-}
-
-const RUNNABLE = ['.sh', '.mjs', '.js', '.yml', '.yaml']
-
-function universe() {
-    const files = []
-    for (const path of trackedFiles()) {
-        if (path.startsWith(EXCLUDED)) continue
-        const full = join(ROOT, path)
-        try {
-            if (!statSync(full).isFile()) continue
-        } catch {
-            continue
-        }
-        let body
-        try {
-            body = readFileSync(full, 'utf8')
-        } catch {
-            continue
-        }
-        const named = RUNNABLE.some((ext) => path.endsWith(ext))
-        /*
-         * A NUL rules out a *candidate* discovered by shebang sniffing, which
-         * has to read every tracked file including binary fixtures. It does not
-         * rule out a file whose extension already says it is a script:
-         * `tests/interop/bindgen-c/check-report.mjs` embeds four NUL bytes in a
-         * fixture string, and skipping it dropped a real Node.js source file
-         * from the census with nothing to show for it.
-         */
-        if (!named && body.indexOf(String.fromCharCode(0)) !== -1) continue
-        if (!named && !body.startsWith('#!')) continue
-        files.push({ path, body })
-    }
-    return files.sort((a, b) => (a.path < b.path ? -1 : 1))
 }
 
 /* One row per (requirement, kind, path), with the number of occurrences and
